make Instance_t unsigned long instead of unsigned short
the DUNE 2x2 and ND uses of larcv produce very large Particle and Interaction IDs (the former because an extremely large number of true particles are simulated; the latter because the interaction ID encodes the GENIE file it came from).  in order to store them successfully without wrapping we promote the basic type to be larger.